+++ src/gems_craft/expression/expression.py
@@ -374,6 +374,16 @@ class ReducedCostNode(ExpressionNode):
     variable_id: str
 
 
+@dataclass(frozen=True, eq=False)
+class LowerBoundNode(ExpressionNode):
+    variable_id: str
+
+
+@dataclass(frozen=True, eq=False)
+class UpperBoundNode(ExpressionNode):
+    variable_id: str
+
+
 def sum_expressions(expressions: Sequence[ExpressionNode]) -> ExpressionNode:
     if len(expressions) == 0:
         return LiteralNode(0)